requirement_gpu.txt
pip==9.0.1
absl-py==0.7.0
astor==0.7.1
certifi
cffi==1.12.2
chardet==3.0.4
Cython==0.29.6
future==0.17.1
gast==0.2.2
glfw==1.7.1
GPUtil==1.4.0
grpcio==1.19.0
gym==0.12.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
json-tricks==3.13.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lockfile==0.12.2
Markdown==3.0.1
mock==2.0.0
numpy==1.16.2
overrides==1.9
pbr==5.1.3
Pillow==5.4.1
protobuf==3.7.0
pycparser==2.19
pyglet==1.3.2
PyOpenGL==3.1.0
requests==2.21.0
scipy==1.2.1
six==1.12.0
tensorflow-gpu==1.12.0
tensorboard==1.12.2
tensorflow-estimator==1.13.0
tensorflow-probability==0.5.0
termcolor==1.1.0
transitions==0.6.9
typeguard==2.1.0
urllib3==1.24.1
Werkzeug==0.14.1
coverage==4.5.2
codecov==2.0.15
pandas==0.24.2
autograd==1.2
gpflow==1.3.0